requirements.txt
websockets
redis>=4.2
motor
asyncio
python-dotenv
httpx
aiohttp
prometheus_client